+++ src/lzma2_reader.rs
@@ -50,16 +50,9 @@ fn get_dict_size(dict_size: u32) -> u32 {
 }
 
 impl<R> LZMA2Reader<R> {
+    /// Unwraps the reader, returning the underlying reader.
     pub fn into_inner(self) -> R {
         self.inner
-    }
-
-    pub fn get_ref(&self) -> &R {
-        &self.inner
-    }
-
-    pub fn get_mut(&mut self) -> &mut R {
-        &mut self.inner
     }
 }